+++ Node.h
@@ -5,7 +5,7 @@ template <typename T>
 class Node {
  private:
   T value;
-  int balance;
+  short int balance;
   Node<T>* leftChild;
   Node<T>* rightChild;
 
@@ -13,8 +13,8 @@ class Node {
   Node<T>(T v);
   T getValue();
 
-  int getBalance();
-  void setBalance(int b);
+  short int getBalance();
+  void setBalance(short int b);
 
   Node<T>*& getLeftChild();
   void setLeftChild(Node<T>& n);
